feat: refactor banners and logic
