+++ plugin-hrm-form/src/components/contact/DraftAndResolvedContactBanners.tsx
@@ -1,0 +1,106 @@
+import React, { useMemo, useState } from 'react';
+import { Close } from '@material-ui/icons';
+import { Template, Manager } from '@twilio/flex-ui';
+import { useDispatch } from 'react-redux';
+
+import { checkTaskAssignment, completeTaskAssignment } from '../../services/ServerlessService';
+import InfoIcon from '../caseMergingBanners/InfoIcon';
+import {
+  HeaderCloseButton,
+  SaveAndEndButton,
+  HiddenText,
+  Flex,
+  BannerContainer,
+  BannerAction,
+  BannerText,
+} from '../../styles';
+import { Contact, RouterTask } from '../../types/types';
+import getCanEditInProgressContact from '../../permissions/canEditInProgressContact';
+import { newFinalizeContactAsyncAction } from '../../states/contacts/saveContact';
+
+type DraftAndResolvedContactBannersProps = {
+  savedContact: Contact;
+  task: RouterTask;
+};
+
+const DraftAndResolvedContactBanners: React.FC<DraftAndResolvedContactBannersProps> = ({ savedContact, task }) => {
+  const [showResolvedBanner, setShowResolvedBanner] = useState(false);
+
+  const dispatch = useDispatch();
+
+  const saveFinalizedContact = (task: RouterTask, contact: Contact) => {
+    dispatch(newFinalizeContactAsyncAction(task, contact));
+  };
+
+  const isDraft = !savedContact.finalizedAt;
+
+  const workerRoles = Manager.getInstance().workerClient.attributes.roles;
+
+  const canEditContact = useMemo(() => getCanEditInProgressContact(savedContact, workerRoles), [
+    savedContact,
+    workerRoles,
+  ]);
+
+  const handleSaveAndEnd = async () => {
+    const status = await checkTaskAssignment(savedContact.taskId);
+
+    const updatedContact = {
+      ...savedContact,
+      rawJson: {
+        ...savedContact.rawJson,
+        callType: 'Uncategorized',
+      },
+    };
+
+    if (status) {
+      // dialog box to confirm if the user wants to save and end the contact
+
+      await completeTaskAssignment(savedContact.taskId);
+    }
+
+    await saveFinalizedContact(task, updatedContact);
+    setShowResolvedBanner(true);
+  };
+
+  return (
+    <>
+      {isDraft && (
+        <BannerContainer color="yellow">
+          <Flex width="100%" alignItems="center">
+            <InfoIcon color="#fed44b" />
+            <BannerText>
+              <Template code="Contact-DraftStatus" />
+            </BannerText>
+            {canEditContact() && (
+              <BannerAction alignRight={true} onClick={handleSaveAndEnd}>
+                <SaveAndEndButton>
+                  <Template code="BottomBar-SaveAndEnd" />
+                </SaveAndEndButton>
+              </BannerAction>
+            )}
+          </Flex>
+        </BannerContainer>
+      )}
+      {showResolvedBanner && (
+        <BannerContainer color="blue" style={{ paddingTop: '12px', paddingBottom: '12px', marginTop: '10px' }}>
+          <Flex width="100%" justifyContent="space-between" alignItems="center">
+            <InfoIcon color="#001489" />
+            <BannerText>
+              <Template code="Contact-ResolvedStatus" />
+            </BannerText>
+            <BannerAction onClick={() => setShowResolvedBanner(false)} alignRight={true} color="black">
+              <HeaderCloseButton />
+              <HiddenText>
+                <Template code="CloseButton" />
+              </HiddenText>
+              <Close fontSize="small" />
+            </BannerAction>
+          </Flex>
+        </BannerContainer>
+      )}
+    </>
+  );
+};
+
+// eslint-disable-next-line import/no-unused-modules
+export default DraftAndResolvedContactBanners;
